flaskBackend/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
filelock==3.15.4
Flask==2.2.5
Flask-APScheduler==1.13.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
flask-restplus==0.13.0
flask-restx==1.3.0
fsspec==2024.6.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
pycparser==2.21
pycryptodome==3.19.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
referencing==0.32.1
rpds-py==0.16.2
six==1.16.0
sympy==1.13.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.1
